refactor prepare

+++ prepare-licheesgnano.sh
@@ -1,5 +1,10 @@
 #!/bin/bash -e
 
+if [ "${SG_BOARD_FAMILY}/${SG_BOARD_LINK}" = "/" ]; then
+  export SG_BOARD_FAMILY=sg200x
+  export SG_BOARD_LINK=sg2002_licheervnano_sd
+fi
+
 if [ ! -e host-tools/gcc ]; then
   bash -e host/replace-all-thead-toolchains.sh
   #bash -e host/replace-all-linaro-toolchains.sh
@@ -7,44 +12,55 @@ if [ ! -e host-tools/gcc ]; then
 fi
 
 if git -C linux_5.10 checkout -b build 2>/dev/null ; then
-  ln -sf ../../../../build/output/sg2002_licheervnano_sd/cvi_board_memmap.h linux_5.10/scripts/dtc/include-prefixes/cvi_board_memmap.h
+  ln -sf ../../../../build/output/${SG_BOARD_LINK}/cvi_board_memmap.h linux_5.10/scripts/dtc/include-prefixes/cvi_board_memmap.h
   git -C linux_5.10 add scripts/dtc/include-prefixes/cvi_board_memmap.h
-  for a in sg200x/sg2000_duo_sd sg200x/sg2002_duo_sd sg200x/sg2002_licheervnano_sd ; do
+  for a in sg200x/sg2000_duo_sd sg200x/sg2002_duo_sd sg200x/sg2002_licheervnano_sd ${SG_BOARD_FAMILY}/${SG_BOARD_LINK} ; do
     d=`pwd`/build/boards/$a/dts_riscv
+    [ -e $d ] || continue
     for f in $d/*.dts ; do
       b=`basename $f`
       ln -sf $d/$b linux_5.10/arch/riscv/boot/dts/cvitek/$b
-      git -C linux_5.10 add arch/riscv/boot/dts/cvitek/$b
+      git -C linux_5.10 add -f arch/riscv/boot/dts/cvitek/$b
     done
   done
-  for a in sg200x/sg2002_licheea53nano_sd sg200x/sg2000_wevb_sd ; do
+  for a in sg200x/sg2002_licheea53nano_sd sg200x/sg2000_wevb_sd ${SG_BOARD_FAMILY}/${SG_BOARD_LINK} ; do
     d=`pwd`/build/boards/$a/dts_arm
+    [ -e $d ] || continue
     for f in $d/*.dts ; do
       b=`basename $f`
       ln -sf $d/$b linux_5.10/arch/arm/boot/dts/cvitek/$b
-      git -C linux_5.10 add arch/arm/boot/dts/cvitek/$b
+      git -C linux_5.10 add -f arch/arm/boot/dts/cvitek/$b
     done
   done
-  d=`pwd`/build/boards/default/dts/sg200x
+  for a in ${SG_BOARD_FAMILY}/${SG_BOARD_LINK} ; do
+    d=`pwd`/build/boards/$a/dts_arm64
+    [ -e $d ] || continue
+    for f in $d/*.dts ; do
+      b=`basename $f`
+      ln -sf $d/$b linux_5.10/arch/arm64/boot/dts/cvitek/$b
+      git -C linux_5.10 add -f arch/arm64/boot/dts/cvitek/$b
+    done
+  done
+  d=`pwd`/build/boards/default/dts/${SG_BOARD_FAMILY}
   for f in $d/*.dtsi ; do
     b=`basename $f`
     ln -sf $d/$b linux_5.10/arch/riscv/boot/dts/cvitek/$b
-    git -C linux_5.10 add arch/riscv/boot/dts/cvitek/$b
+    git -C linux_5.10 add -f arch/riscv/boot/dts/cvitek/$b
     ln -sf $d/$b linux_5.10/arch/arm/boot/dts/cvitek/$b
-    git -C linux_5.10 add arch/arm/boot/dts/cvitek/$b
+    git -C linux_5.10 add -f arch/arm/boot/dts/cvitek/$b
   done
-  git -C linux_5.10 commit -m "licheervnano: symlink dts and memmap"
+  git -C linux_5.10 commit -m "licheesgnano: symlink dts and memmap"
 fi
 
 if git -C u-boot-2021.10 checkout -b build 2>/dev/null ; then
-  d=`pwd`/build/boards/sg200x/sg2002_licheervnano_sd/u-boot
+  d=`pwd`/build/boards/${SG_BOARD_FAMILY}/${SG_BOARD_LINK}/u-boot
   ln -sf $d/cvi_board_init.c u-boot-2021.10/board/cvitek/cvi_board_init.c
   git -C u-boot-2021.10 add board/cvitek/cvi_board_init.c
   ln -sf $d/cvitek.h u-boot-2021.10/include/cvitek/cvitek.h
   git -C u-boot-2021.10 add include/cvitek/cvitek.h
-  ln -sf ../../build/output/sg2002_licheervnano_sd/cvi_board_memmap.h u-boot-2021.10/include/cvi_board_memmap.h
+  ln -sf ../../build/output/${SG_BOARD_LINK}/cvi_board_memmap.h u-boot-2021.10/include/cvi_board_memmap.h
   git -C u-boot-2021.10 add include/cvi_board_memmap.h
-  git -C u-boot-2021.10 commit -m "licheervnano: symlink dts and memmap"
+  git -C u-boot-2021.10 commit -m "licheesgnano: symlink dts and memmap"
 fi
 
 KERNEL_PATH=linux_5.10
